TS-42/TS-117: more int/ptr size conflicts.  My compilers don't produce an error, but these are some fixes the additional cases I could find.
git-svn-id: https://svn.apache.org/repos/asf/incubator/trafficserver/traffic/trunk@905519 13f79535-47bb-0310-9956-ffa450edef68

+++ iocore/cache/CacheRead.cc
@@ -170,7 +170,7 @@ CacheVC::openReadChooseWriter(int event, Event * e)
   NOWARN_UNUSED(e);
   NOWARN_UNUSED(event);
 
-  int err = ECACHE_DOC_BUSY;
+  intptr_t err = ECACHE_DOC_BUSY;
   CacheVC *w = NULL;
 
   if (!f.http_request) {
@@ -182,7 +182,7 @@ CacheVC::openReadChooseWriter(int event, Event * e)
       return openReadStartHead(EVENT_IMMEDIATE, 0);
     }
     if (!w->closed) {
-      return openReadFromWriterFailure(CACHE_EVENT_OPEN_READ_FAILED, (Event *) - err);
+      return openReadFromWriterFailure(CACHE_EVENT_OPEN_READ_FAILED, (Event *) -err);
     }
     write_vc = w;
   }
@@ -285,10 +285,10 @@ CacheVC::openReadFromWriter(int event, Event * e)
     f.read_from_writer_called = 1;
   }
   cancel_trigger();
-  int err = ECACHE_DOC_BUSY;
+  intptr_t err = ECACHE_DOC_BUSY;
   Debug("cache_read_agg", "%x: key: %X In openReadFromWriter", this, first_key.word(1));
 #ifndef READ_WHILE_WRITER
-  return openReadFromWriterFailure(CACHE_EVENT_OPEN_READ_FAILED, (Event *) - err);
+  return openReadFromWriterFailure(CACHE_EVENT_OPEN_READ_FAILED, (Event *) -err);
 #else
   if (_action.cancelled)
     return free_CacheVC(this);
@@ -697,7 +697,7 @@ CacheVC::openReadStartEarliest(int event, Event * e)
   NOWARN_UNUSED(e);
   NOWARN_UNUSED(event);
 
-  int err = ECACHE_NO_DOC;
+  intptr_t err = ECACHE_NO_DOC;
   Doc *doc = NULL;
 
   cancel_trigger();
@@ -841,7 +841,7 @@ CacheVC::openReadVecWrite(int event, Event * e)
   NOWARN_UNUSED(event);
 
   ink_debug_assert(part->mutex->thread_holding == this_ethread());
-  int err = ECACHE_ALT_MISS;
+  intptr_t err = ECACHE_ALT_MISS;
   ink_assert(event == AIO_EVENT_DONE);
   set_io_not_in_progress();
   ink_assert(od);
@@ -890,7 +890,7 @@ CacheVC::openReadVecWrite(int event, Event * e)
 int
 CacheVC::openReadStartHead(int event, Event * e)
 {
-  int err = ECACHE_NO_DOC;
+  intptr_t err = ECACHE_NO_DOC;
   Doc *doc = NULL;
   cancel_trigger();
   set_io_not_in_progress();
